If we don't get up to now, at least get recent
When we exceed the limits on MAM fetching for catchup, we will have the
750 messages that came after what we last received, but no recent
messages.  Toggling the account off and on again will get the next 750
and so on until we get up until now, or if we get a live message that
will be used as the new sync point. However this can be very confusing
to a user who sees only messages from weeks ago (if they have been
offline for quite some time) and nothing recent, even though there is
recent activity.

So detect when we exceed a limit and grab some recent activity also to
make things at least look natural.  This does result in a hole in the
MAM sync state, but so did the old way eventually, and this at least is
less confusing.